Replace bicycle point with vec2
